requirements.txt
dtlpy
pyyaml>=5.3.1
onnxruntime
pillow>=11.0.0
torch
transformers
opencv-python
requests
tqdm
scipy